+++ .gitignore
@@ -2,5 +2,4 @@
 *.sw[po]
 
 node_modules/
-mounted-volume
-input/*
+mounted-volume/
