+++ example-volume/input/table-of-contents/layout.css
@@ -1,6 +1,7 @@
 body {
     font-size: 12px;
     margin: 0;
+    color: red;
 }
 
 img {
