--- src/mount/config.yml
@@ -1,18 +1,0 @@
-name: 'Cats and Dogs'
-size:
-  width: '11in'
-  height: '8.5in'
-pages:
-  page:
-    layout_table:
-      html: 'input/table-of-contents/layout.html'
-      css: 'input/table-of-contents/layout.css'
-  page_sequence:
-    layout_image:
-      html: 'input/image-page-sequence/1-image-color/layout.html'
-      css: 'input/image-page-sequence/1-image-color/layout.css'
-    layout_notes:
-      html: 'input/image-page-sequence/2-image-notes/layout.html'
-      css: 'input/image-page-sequence/2-image-notes/layout.css'
-    media: 'input/image-page-sequence/media'
-data: 'input/data.json'
