--- src/mount/input/data.json
@@ -1,26 +1,0 @@
-{
-    "playlists": [{
-            "playlist_name": "Cats",
-            "images": [{
-                "image_name": "Cool Cat",
-                "image_id": "cat1"
-            }, {
-                "image_name": "Cooler Cat",
-                "image_id": "cat2"
-            }, {
-                "image_name": "Coolest Cat",
-                "image_id": "cat3"
-            }]
-        },
-        {
-            "playlist_name": "Dogs",
-            "images": [{
-                "image_name": "Cool Dog",
-                "image_id": "dog1"
-            }, {
-                "image_name": "Cooler Dog",
-                "image_id": "dog2"
-            }]
-        }
-    ]
-}
