--- src/mount/input/image-page-sequence/1-image-color/layout.css
@@ -1,4 +1,0 @@
-.image {
-    text-align: center;
-    padding: 15% 0;
-}
