--- src/mount/input/image-page-sequence/1-image-color/layout.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-</head>
-
-<body id="root">
-    <div class="page-dim">
-
-        <div class="image">
-            <div id="image" />
-        </div>
-
-        <div class="footer">
-            <div id="page-count" />
-        </div>
-
-    </div>
-</body>
-
-</html>
