--- src/mount/input/image-page-sequence/2-image-notes/layout.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-</head>
-
-<body id="root">
-    <div class=page-dim>
-        <div class="footer">
-            <div id="page-count" />
-        </div>
-    </div>
-
-</body>
-
-</html>
