--- src/mount/input/table-of-contents/layout.css
@@ -1,44 +1,0 @@
-body {
-    font-size: 12px;
-    margin: 0;
-}
-
-img {
-    width: 80%;
-    height: auto;
-}
-
-.footer {
-    position: absolute;
-    bottom: 0;
-    width: 100%;
-    height: 50px;
-    border-top: solid black 1px;
-}
-
-.playlist-container {
-    padding: 0px 30px 15px 30px;
-}
-
-.playlist {
-    display: grid;
-    grid-column-gap: 30px;
-    grid-template-columns: 33.3% auto 33.3%;
-    padding-bottom: 15px;
-
-}
-
-.playlist>div {
-    background-color: rgba(255, 255, 255, 0.8);
-    text-align: center;
-}
-
-.playlist-title {
-    font-weight: bold;
-    font-size: 16px;
-}
-
-.item {
-    display: flex;
-    justify-content: space-between;
-}
