--- src/mount/input/table-of-contents/layout.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-</head>
-
-<body id="root">
-    <div class=page-dim>
-        <h1>Table of Contents</h1>
-        <div id="playlist" />
-    </div>
-
-</body>
-
-</html>
